Add files via upload
This is a research proposal I wrote in March 2023 for the Ph.D. researcher position on nudging active mobility. There I came up with the idea for the field experiment on cooperation between a local bike-sharing company and the Norwegian municipality. The main idea was to imply so-called gamification and sustainability framing techniques. Since this proposal didn't get any continuation, everyone is free to use presented ideas and pieces of information for there on research and work projects related to promoting active ways of urban transportation or for any other topics related to behavioral economics.